Refactor test_runner
